requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
kafka-python==2.0.2
pandas==2.1.3
numpy==1.24.4
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
shap==0.43.0
lime==0.2.0.1
typer==0.9.0
rich==13.7.0
imbalanced-learn==0.11.0
joblib==1.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
